OCR_and_Classify_with_Qwen2VL/requirements.txt
torch>=2.2.0
torchvision>=0.16.0
transformers>=4.34.0
Pillow>=10.0.0
numpy>=1.25.0
gradio>=3.53.1
accelerate>=0.23.0
datasets>=3.0.1
